Design: Reflect changed announcement design
